requirements.txt
pandas==2.2.2
SQLAlchemy==2.0.32
python-dateutil==2.9.0.post0
apache-airflow==2.9.2
pytest==8.3.2
